api: add u32 precision type

+++ src/cpu/reference_pooling.cpp
@@ -108,7 +108,7 @@ status_t reference_pooling<prec>::primitive_desc_init(
         &pool_d.src_desc.tensor_desc, f32, nchw));
     if (pool_d.indices_desc.format == any)
         CHECK(mkl_dnn_memory_desc_init(&pool_d.indices_desc,
-        &pool_d.indices_desc.tensor_desc, f32,nchw));
+        &pool_d.indices_desc.tensor_desc, u32, nchw));
     if (pool_d.dst_desc.format == any)
         CHECK(mkl_dnn_memory_desc_init(&pool_d.dst_desc,
         &pool_d.dst_desc.tensor_desc, f32, nchw));
